Move impl FromStr for Netv4Addr into netv4addr mod
Tested-by: Kristofer Rye <kristofer.rye@gmail.com>
